Roll the testing service account
The new service account belongs to a new GCP project, separate from the gargle GCP project.

This means that googlesheets4 dev work, such as GHA jobs, is not associated with the same GCP project that owns the tidyverse OAuth client. This means dev work won't be competing with regular users for project-level quota.

[covr]