Remove the push trigger for pkgdown workflow
Previously, this was rigged so that a commit message including '[pkgdown]', on main, would trigger the workflow.

But I think the workflow dispatch trigger basically meets this need now and lets me simplify this workflow.

As it was, the release trigger was broken, since it wasn't included specifically in the `if:`, which was made necessary by the special commit message.